roles and profiles: admin/foreman, reposerver/foreman
Just to make some names differet to see how things get put together,
we use role admin which installs the profile foreman.

katello is basically foreman, but with added repository management,
so we call the role reposerver, which also installs the profile
foreman (to begin with).

Just to get us started, we're keeping it simple by just installing
the foreman repository.